Patches boolean json unmarshall error on slack service
